Enum mpl_nft_packs::error::NFTPacksError

pub enum NFTPacksError {

    WrongAllowedAmountToRedeem,
    WrongRedeemDate,
    CardProbabilityMissing,
    WrongCardProbability,
    CardShouldntHaveProbabilityValue,
    ProvedVouchersMismatchPackVouchers,
    PackIsAlreadyEnded,
    PackSetNotConfigured,
    CantActivatePack,
    PackSetNotActivated,
    ProvingPackProcessCompleted,
    ProvingVoucherProcessCompleted,
    WrongEdition,
    WrongEditionMint,
    Overflow,
    Underflow,
    NotEmptyPackSet,
    WrongPackState,
    ImmutablePackSet,
    CantSetTheSameValue,
    WrongMaxSupply,
    WrongVoucherSupply,
    RandomOracleOutOfDate,
    CardDoesntHaveEditions,
    UserRedeemedAllCards,
    UriTooLong,
    CardDoesntHaveMaxSupply,
    WrongMasterSupply,
    MissingEditionsInPack,
    AlreadySetNextCardToRedeem,
    EndDateNotArrived,
    DescriptionTooLong,
    WhitelistedCreatorInactive,
    WrongWhitelistedCreator,
    WrongVoucherOwner,
    CardShouldntHaveSupplyValue,
    PackIsFullWithCards,
    WeightsNotCleanedUp,
    CardAlreadyRedeemed,
    UserCantRedeemThisCard,
    InvalidWeightPosition,

}

Errors that may be returned by the Metaplex NFT packs program.

Variants

WrongAllowedAmountToRedeem

Allowed amount to redeem should be more then 0

WrongRedeemDate

Wrong redeem date

CardProbabilityMissing

Card probability is missing

WrongCardProbability

Wrong card probability value

CardShouldntHaveProbabilityValue

Cards for this pack shouldn’t have probability value

ProvedVouchersMismatchPackVouchers

Proved vouchers mismatch pack vouchers

PackIsAlreadyEnded

Pack is already ended

PackSetNotConfigured

NFT pack set not fully configured

CantActivatePack

Can’t activate NFT pack in current state

PackSetNotActivated

Pack set should be activated

ProvingPackProcessCompleted

Proving process for this pack is completed

ProvingVoucherProcessCompleted

Proving process for this voucher is completed

WrongEdition

Received edition from wrong master

WrongEditionMint

Received wrong edition mint

Overflow

Overflow

Underflow

Underflow

NotEmptyPackSet

Pack set should be empty to delete it

WrongPackState

Wrong pack state to change data

ImmutablePackSet

Pack set is immutable

CantSetTheSameValue

Can’t set the same value

WrongMaxSupply

Wrong max supply value

WrongVoucherSupply

Voucher should have supply greater then 0

RandomOracleOutOfDate

Random oracle updated long time ago

CardDoesntHaveEditions

Card ran out of editions

UserRedeemedAllCards

User redeemed all allowed cards

UriTooLong

URI too long

CardDoesntHaveMaxSupply

Card doesn’t have max supply

WrongMasterSupply

Master edition should have unlimited supply

MissingEditionsInPack

Pack set doesn’t have total editions

AlreadySetNextCardToRedeem

User already got next card to redeem

EndDateNotArrived

Can’t close the pack before end date

DescriptionTooLong

Pack description too long

WhitelistedCreatorInactive

Whitelisted creator inactive

WrongWhitelistedCreator

Wrong whitelisted creator address

WrongVoucherOwner

Voucher owner mismatch

CardShouldntHaveSupplyValue

Cards for this pack shouldn’t have supply value

PackIsFullWithCards

Pack is already full of cards

WeightsNotCleanedUp

Card weights should be cleaned up

CardAlreadyRedeemed

User already redeemed this card

UserCantRedeemThisCard

User can’t redeem this card

InvalidWeightPosition

Invalid weight position